renamed to .cjs
